Refactor codegen to allow calling functions without arguments as an expression
