+++ codegen/decl.go
@@ -10,12 +10,7 @@ import (
 	"github.com/pkg/errors"
 )
 
-func (cg *CodeGen) EmitFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, call *parser.CallStmt, ac aliasCallback, chainStart interface{}) (interface{}, error) {
-	var args []*parser.Expr
-	if call != nil {
-		args = call.Args
-	}
-
+func (cg *CodeGen) EmitFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, args []*parser.Expr, ac aliasCallback, chainStart interface{}) (interface{}, error) {
 	nonVariadicArgs := 0
 	for _, field := range fun.Params.List {
 		if field.Variadic == nil {
@@ -26,66 +21,54 @@ func (cg *CodeGen) EmitFuncDecl(ctx context.Context, scope *parser.Scope, fun *p
 		return nil, errors.WithStack(errors.Errorf("%s expected args %s, found %s", fun.Name, fun.Params.List, args))
 	}
 
-	err := cg.ParameterizedScope(ctx, scope, call, fun, args, ac, chainStart)
+	err := cg.ParameterizedScope(ctx, scope, fun, args, ac)
 	if err != nil {
 		return nil, err
 	}
 
-	v := chainStart
-	switch fun.Type.Primary() {
-	case parser.Filesystem, parser.Option:
-		if _, ok := v.(llb.State); v == nil || !ok {
-			v = llb.Scratch()
-		}
-	case parser.Str:
-		if _, ok := v.(string); v == nil || !ok {
-			v = ""
-		}
-	}
-
 	// Before executing a function.
-	err = cg.Debug(ctx, fun.Scope, fun, v)
+	err = cg.Debug(ctx, fun.Scope, fun, chainStart)
 	if err != nil {
-		return v, err
+		return chainStart, err
 	}
 
 	switch fun.Type.Primary() {
 	case parser.Filesystem:
-		return cg.EmitFilesystemBlock(ctx, fun.Scope, fun.Body.NonEmptyStmts(), ac, v)
+		return cg.EmitFilesystemBlock(ctx, fun.Scope, fun.Body.NonEmptyStmts(), ac, chainStart)
 	case parser.Option:
 		return cg.EmitOptions(ctx, fun.Scope, string(fun.Type.Secondary()), fun.Body.NonEmptyStmts(), ac)
 	case parser.Str:
-		return cg.EmitStringBlock(ctx, fun.Scope, fun.Body.NonEmptyStmts(), v)
+		return cg.EmitStringBlock(ctx, fun.Scope, fun.Body.NonEmptyStmts(), chainStart)
 	case parser.Group:
-		return cg.EmitGroupBlock(ctx, fun.Scope, fun.Body.NonEmptyStmts(), ac, v)
+		return cg.EmitGroupBlock(ctx, fun.Scope, fun.Body.NonEmptyStmts(), ac, chainStart)
 	default:
-		return v, checker.ErrInvalidTarget{Node: fun}
+		return chainStart, checker.ErrInvalidTarget{Node: fun}
 	}
 }
 
-func (cg *CodeGen) EmitFilesystemFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, call *parser.CallStmt, ac aliasCallback, chainStart interface{}) (llb.State, error) {
-	v, err := cg.EmitFuncDecl(ctx, scope, fun, call, ac, chainStart)
+func (cg *CodeGen) EmitFilesystemFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, args []*parser.Expr, ac aliasCallback, chainStart interface{}) (llb.State, error) {
+	v, err := cg.EmitFuncDecl(ctx, scope, fun, args, ac, chainStart)
 	return v.(llb.State), err
 }
 
-func (cg *CodeGen) EmitOptionFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, call *parser.CallStmt) ([]interface{}, error) {
-	v, err := cg.EmitFuncDecl(ctx, scope, fun, call, noopAliasCallback, nil)
+func (cg *CodeGen) EmitOptionFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, args []*parser.Expr) ([]interface{}, error) {
+	v, err := cg.EmitFuncDecl(ctx, scope, fun, args, noopAliasCallback, nil)
 	if v == nil {
 		return nil, err
 	}
 	return v.([]interface{}), err
 }
 
-func (cg *CodeGen) EmitStringFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, call *parser.CallStmt, ac aliasCallback, chainStart interface{}) (string, error) {
-	v, err := cg.EmitFuncDecl(ctx, scope, fun, call, ac, chainStart)
+func (cg *CodeGen) EmitStringFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, args []*parser.Expr, ac aliasCallback, chainStart interface{}) (string, error) {
+	v, err := cg.EmitFuncDecl(ctx, scope, fun, args, ac, chainStart)
 	if v == nil {
 		return "", err
 	}
 	return v.(string), err
 }
 
-func (cg *CodeGen) EmitGroupFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, call *parser.CallStmt, ac aliasCallback, chainStart interface{}) (solver.Request, error) {
-	v, err := cg.EmitFuncDecl(ctx, scope, fun, call, ac, chainStart)
+func (cg *CodeGen) EmitGroupFuncDecl(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, args []*parser.Expr, ac aliasCallback, chainStart interface{}) (solver.Request, error) {
+	v, err := cg.EmitFuncDecl(ctx, scope, fun, args, ac, chainStart)
 	if v == nil {
 		return nil, err
 	}
@@ -112,9 +95,9 @@ func (cg *CodeGen) EmitGroupFuncDecl(ctx context.Context, scope *parser.Scope, f
 	return request, err
 }
 
-func (cg *CodeGen) EmitAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, call *parser.CallStmt, chainStart interface{}) (interface{}, error) {
+func (cg *CodeGen) EmitAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, args []*parser.Expr, chainStart interface{}) (interface{}, error) {
 	var v interface{}
-	_, err := cg.EmitFuncDecl(ctx, scope, alias.Func, call, func(aliasCall *parser.CallStmt, aliasValue interface{}) bool {
+	_, err := cg.EmitFuncDecl(ctx, scope, alias.Func, args, func(aliasCall *parser.CallStmt, aliasValue interface{}) bool {
 		if alias.Call == aliasCall {
 			v = aliasValue
 			return false
@@ -127,31 +110,31 @@ func (cg *CodeGen) EmitAliasDecl(ctx context.Context, scope *parser.Scope, alias
 	return v, err
 }
 
-func (cg *CodeGen) EmitFilesystemAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, call *parser.CallStmt, chainStart interface{}) (llb.State, error) {
-	v, err := cg.EmitAliasDecl(ctx, scope, alias, call, chainStart)
+func (cg *CodeGen) EmitFilesystemAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, args []*parser.Expr, chainStart interface{}) (llb.State, error) {
+	v, err := cg.EmitAliasDecl(ctx, scope, alias, args, chainStart)
 	if v == nil {
 		return llb.Scratch(), err
 	}
 	return v.(llb.State), err
 }
 
-func (cg *CodeGen) EmitStringAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, call *parser.CallStmt, chainStart interface{}) (string, error) {
-	v, err := cg.EmitAliasDecl(ctx, scope, alias, call, chainStart)
+func (cg *CodeGen) EmitStringAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, args []*parser.Expr, chainStart interface{}) (string, error) {
+	v, err := cg.EmitAliasDecl(ctx, scope, alias, args, chainStart)
 	if v == nil {
 		return "", err
 	}
 	return v.(string), err
 }
 
-func (cg *CodeGen) EmitGroupAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, call *parser.CallStmt, chainStart interface{}) (solver.Request, error) {
-	v, err := cg.EmitAliasDecl(ctx, scope, alias, call, chainStart)
+func (cg *CodeGen) EmitGroupAliasDecl(ctx context.Context, scope *parser.Scope, alias *parser.AliasDecl, args []*parser.Expr, chainStart interface{}) (solver.Request, error) {
+	v, err := cg.EmitAliasDecl(ctx, scope, alias, args, chainStart)
 	if v == nil {
 		return nil, err
 	}
 	return v.(solver.Request), err
 }
 
-func (cg *CodeGen) ParameterizedScope(ctx context.Context, scope *parser.Scope, call *parser.CallStmt, fun *parser.FuncDecl, args []*parser.Expr, ac aliasCallback, chainStart interface{}) error {
+func (cg *CodeGen) ParameterizedScope(ctx context.Context, scope *parser.Scope, fun *parser.FuncDecl, args []*parser.Expr, ac aliasCallback) error {
 	for i, field := range fun.Params.List {
 		var (
 			data interface{}
@@ -162,7 +145,7 @@ func (cg *CodeGen) ParameterizedScope(ctx context.Context, scope *parser.Scope, 
 		switch typ {
 		case parser.Str:
 			var v string
-			v, err = cg.EmitStringExpr(ctx, scope, call, args[i])
+			v, err = cg.EmitStringExpr(ctx, scope, args[i])
 			data = v
 		case parser.Int:
 			var v int
@@ -174,7 +157,7 @@ func (cg *CodeGen) ParameterizedScope(ctx context.Context, scope *parser.Scope, 
 			data = v
 		case parser.Filesystem:
 			var v llb.State
-			v, err = cg.EmitFilesystemExpr(ctx, scope, nil, args[i], ac, chainStart)
+			v, err = cg.EmitFilesystemExpr(ctx, scope, args[i], ac)
 			data = v
 		case parser.Option:
 			var v []interface{}
@@ -193,11 +176,11 @@ func (cg *CodeGen) ParameterizedScope(ctx context.Context, scope *parser.Scope, 
 			data = v
 		case parser.Group:
 			var v solver.Request
-			v, err = cg.EmitGroupExpr(ctx, scope, nil, args[i], ac, chainStart)
+			v, err = cg.EmitGroupExpr(ctx, scope, args[i], ac, nil)
 			data = v
 		}
 		if err != nil {
-			return ErrCodeGen{Node: call, Err: err}
+			return ErrCodeGen{Node: args[i], Err: err}
 		}
 
 		fun.Scope.Insert(&parser.Object{
